+++ test/unit/server/get_events_test.js
@@ -1,8 +1,8 @@
-const MockAdapter = require('axios-mock-adapter');
+const { Server, AxiosClient, nativeToScVal } = SorobanClient;
 
-describe('Server#getEvents', function () {
+describe("Server#getEvents", function () {
   beforeEach(function () {
-    this.server = new SorobanClient.Server(serverUrl);
+    this.server = new Server(serverUrl);
     this.axiosMock = sinon.mock(AxiosClient);
   });
 
@@ -11,16 +11,16 @@ describe('Server#getEvents', function () {
     this.axiosMock.restore();
   });
 
-  it('requests the correct endpoint', function (done) {
+  it("requests the correct endpoint", function (done) {
     let result = { latestLedger: 0, events: [] };
     setupMock(
       this.axiosMock,
       {
         filters: [],
         pagination: {},
-        startLedger: '1'
+        startLedger: "1",
       },
-      result
+      result,
     );
 
     this.server
@@ -34,24 +34,25 @@ describe('Server#getEvents', function () {
       });
   });
 
-  it('can build wildcard filters', function (done) {
+  it("can build wildcard filters", function (done) {
     let result = {
       latestLedger: 1,
-      events: filterEvents(getEventsResponseFixture, '*/*')
+      events: filterEvents(getEventsResponseFixture, "*/*"),
     };
+    expect(result.events).to.not.have.lengthOf(0, JSON.stringify(result));
 
     setupMock(
       this.axiosMock,
       {
-        startLedger: '1',
+        startLedger: "1",
         filters: [
           {
-            topics: [['*', '*']]
-          }
+            topics: [["*", "*"]],
+          },
         ],
-        pagination: {}
+        pagination: {},
       },
-      result
+      result,
     );
 
     this.server
@@ -59,9 +60,9 @@ describe('Server#getEvents', function () {
         startLedger: 1,
         filters: [
           {
-            topics: [['*', '*']]
-          }
-        ]
+            topics: [["*", "*"]],
+          },
+        ],
       })
       .then(function (response) {
         expect(response).to.be.deep.equal(parseEvents(result));
@@ -70,27 +71,28 @@ describe('Server#getEvents', function () {
       .catch(done);
   });
 
-  it('can build matching filters', function (done) {
+  it("can build matching filters", function (done) {
     let result = {
       latestLedger: 1,
       events: filterEvents(
         getEventsResponseFixture,
-        'AAAABQAAAAh0cmFuc2Zlcg==/AAAAAQB6Mcc='
-      )
+        `${topicVals[0]}/${topicVals[1]}`,
+      ),
     };
+    expect(result.events).to.not.have.lengthOf(0, JSON.stringify(result));
 
     setupMock(
       this.axiosMock,
       {
-        startLedger: '1',
+        startLedger: "1",
         filters: [
           {
-            topics: [['AAAABQAAAAh0cmFuc2Zlcg==', 'AAAAAQB6Mcc=']]
-          }
+            topics: [[topicVals[0], topicVals[1]]],
+          },
         ],
-        pagination: {}
+        pagination: {},
       },
-      result
+      result,
     );
 
     this.server
@@ -98,93 +100,95 @@ describe('Server#getEvents', function () {
         startLedger: 1,
         filters: [
           {
-            topics: [['AAAABQAAAAh0cmFuc2Zlcg==', 'AAAAAQB6Mcc=']]
-          }
-        ]
+            topics: [[topicVals[0], topicVals[1]]],
+          },
+        ],
       })
       .then(function (response) {
-        expect(response).to.be.deep.equal(result);
+        expect(response).to.be.deep.equal(parseEvents(result));
         done();
       })
       .catch(done);
   });
 
-  it('can build mixed filters', function (done) {
+  it("can build mixed filters", function (done) {
     let result = {
-      latestLedger: 1,
+      latestLedger: 3,
       events: filterEventsByLedger(
-        filterEvents(getEventsResponseFixture, 'AAAABQAAAAh0cmFuc2Zlcg==/*'),
-        1
-      )
+        filterEvents(getEventsResponseFixture, `${topicVals[0]}/*`),
+        2,
+      ),
     };
+    expect(result.events).to.not.have.lengthOf(0, JSON.stringify(result));
 
     setupMock(
       this.axiosMock,
       {
-        startLedger: '1',
+        startLedger: "2",
         filters: [
           {
-            topics: [['AAAABQAAAAh0cmFuc2Zlcg==', '*']]
-          }
+            topics: [[topicVals[0], "*"]],
+          },
         ],
-        pagination: {}
+        pagination: {},
       },
-      result
+      result,
     );
 
     this.server
       .getEvents({
-        startLedger: 1,
+        startLedger: 2,
         filters: [
           {
-            topics: [['AAAABQAAAAh0cmFuc2Zlcg==', '*']]
-          }
-        ]
+            topics: [[topicVals[0], "*"]],
+          },
+        ],
       })
       .then(function (response) {
-        expect(response).to.be.deep.equal(result);
+        expect(response).to.be.deep.equal(parseEvents(result));
         done();
       })
       .catch(done);
   });
 
-  it('can paginate', function (done) {
+  it("can paginate", function (done) {
     let result = {
-      latestLedger: 1,
+      latestLedger: 3,
       events: filterEventsByLedger(
-        filterEvents(getEventsResponseFixture, '*/*'),
-        1
-      )
+        filterEvents(getEventsResponseFixture, "*/*"),
+        2,
+      ),
     };
+    expect(result.events).to.not.have.lengthOf(0, JSON.stringify(result));
 
     setupMock(
       this.axiosMock,
       {
         filters: [
           {
-            topics: [['*', '*']]
-          }
+            topics: [["*", "*"]],
+          },
         ],
         pagination: {
           limit: 10,
-          cursor: '0164090849041387521-0000000000'
-        }
+          cursor: "0164090849041387521-0000000000",
+        },
       },
-      result
+      result,
     );
 
     this.server
       .getEvents({
         filters: [
           {
-            topics: [['*', '*']]
-          }
+            topics: [["*", "*"]],
+          },
         ],
-        cursor: '0164090849041387521-0000000000',
-        limit: 10
+        cursor: "0164090849041387521-0000000000",
+        limit: 10,
       })
       .then(function (response) {
-        expect(response).to.be.deep.equal(result);
+        expect(response).to.be.deep.equal(parseEvents(result));
         done();
       })
       .catch(done);
@@ -192,93 +196,94 @@ describe('Server#getEvents', function () {
 });
 
 function filterEvents(events, filter) {
+  const parts = filter.split("/");
   return events.filter(
     (e, i) =>
-      e.topic.length == filter.length &&
-      e.topic.every((s, j) => s === filter[j] || s === '*')
+      e.topic.length == parts.length &&
+      e.topic.every((s, j) => s === parts[j] || parts[j] === "*"),
   );
 }
 
 function filterEventsByLedger(events, start) {
   return events.filter((e) => {
-    return e.ledger.parseInt() >= start;
+    return parseInt(e.ledger) >= start;
   });
 }
 
 function setupMock(axiosMock, params, result) {
   axiosMock
-    .expects('post')
+    .expects("post")
     .withArgs(serverUrl, {
-      jsonrpc: '2.0',
+      jsonrpc: "2.0",
       id: 1,
-      method: 'getEvents',
-      params: params
+      method: "getEvents",
+      params: params,
     })
     .returns(Promise.resolve({ data: { result } }));
 }
 
 function parseEvents(result) {
-  return {
-    ...result,
-    events: result.events.map(SorobanClient.parseRawEvents)
-  };
+  return SorobanClient.parseRawEvents(result);
 }
 
+const contractId = "CA3D5KRYM6CB7OWQ6TWYRR3Z4T7GNZLKERYNZGGA5SOAOPIFY6YQGAXE";
+const topicVals = [
+  nativeToScVal("transfer", { type: "symbol" }).toXDR("base64"),
+  nativeToScVal(contractId, { type: "address" }).toXDR("base64"),
+  nativeToScVal(1234).toXDR("base64"),
+];
+let eventVal = nativeToScVal("wassup").toXDR("base64");
 let getEventsResponseFixture = [
   {
-    type: 'system',
-    ledger: '1',
-    ledgerClosedAt: '2022-11-16T16:10:41Z',
-    contractId:
-      'e3e82a76cc316f6289fd1ffbdf315da0f2c6be9582b84b9983a402f02ea0fff7',
-    id: '0164090849041387521-0000000003',
-    pagingToken: '164090849041387521-3',
-    topic: ['AAAABQAAAAh0cmFuc2Zlcg==', 'AAAAAQB6Mcc='],
+    type: "system",
+    ledger: "1",
+    ledgerClosedAt: "2022-11-16T16:10:41Z",
+    contractId,
+    id: "0164090849041387521-0000000003",
+    pagingToken: "164090849041387521-3",
     inSuccessfulContractCall: true,
+    topic: topicVals.slice(0, 2),
     value: {
-      xdr: 'AAAABQAAAApHaWJNb255UGxzAAA='
-    }
+      xdr: eventVal,
+    },
   },
   {
-    type: 'contract',
-    ledger: '2',
-    ledgerClosedAt: '2022-11-16T16:10:41Z',
-    contractId:
-      'e3e82a76cc316f6289fd1ffbdf315da0f2c6be9582b84b9983a402f02ea0fff7',
-    id: '0164090849041387521-0000000003',
-    pagingToken: '164090849041387521-3',
-    topic: ['AAAAAQB6Mcc=', 'AAAABQAAAAh0cmFuc2Zlcg=='],
+    type: "contract",
+    ledger: "2",
+    ledgerClosedAt: "2022-11-16T16:10:41Z",
+    contractId,
+    id: "0164090849041387521-0000000003",
+    pagingToken: "164090849041387521-3",
     inSuccessfulContractCall: true,
+    topic: topicVals.slice(0, 2),
     value: {
-      xdr: 'AAAABQAAAApHaWJNb255UGxzAAA='
-    }
+      xdr: eventVal,
+    },
   },
   {
-    type: 'diagnostic',
-    ledger: '2',
-    ledgerClosedAt: '2022-11-16T16:10:41Z',
-    contractId:
-      'a3e82a76cc316f6289fd1ffbdf315da0f2c6be9582b84b9983a402f02ea0fff7',
-    id: '0164090849041387521-0000000003',
-    pagingToken: '164090849041387521-3',
+    type: "diagnostic",
+    ledger: "2",
+    ledgerClosedAt: "2022-11-16T16:10:41Z",
+    contractId,
+    id: "0164090849041387521-0000000003",
+    pagingToken: "164090849041387521-3",
     inSuccessfulContractCall: true,
-    topic: ['AAAAAQB6Mcc='],
+    topic: [topicVals[0]],
     value: {
-      xdr: 'AAAABQAAAApHaWJNb255UGxzAAA='
-    }
+      xdr: eventVal,
+    },
   },
   {
-    type: 'contract',
-    ledger: '3',
-    ledgerClosedAt: '2022-12-14T01:01:20Z',
-    contractId:
-      '6ebe0114ae15f72f187f05d06dcb66b22bd97218755c9b4646b034ab961fc1d5',
-    id: '0000000171798695936-0000000001',
-    pagingToken: '0000000171798695936-0000000001',
+    type: "contract",
+    ledger: "3",
+    ledgerClosedAt: "2022-12-14T01:01:20Z",
+    contractId,
+    id: "0000000171798695936-0000000001",
+    pagingToken: "0000000171798695936-0000000001",
     inSuccessfulContractCall: true,
-    topic: ['AAAABQAAAAdDT1VOVEVSAA==', 'AAAABQAAAAlpbmNyZW1lbnQAAAA='],
+    topic: topicVals,
     value: {
-      xdr: 'AAAAAQAAAAE='
-    }
-  }
+      xdr: eventVal,
+    },
+  },
 ];