+++ CHANGELOG.md
@@ -5,6 +5,13 @@
 A breaking change should be clearly marked in this log.
 
 
+## Unreleased
+
+### Fixed
+* The `Server.getEvents` method now correctly parses responses without a `contractId` field set: the `events[i].contractId` field on an event will be omitted if there was no ID in the event (e.g. for system events; ([#TODO](https://github.com/stellar/js-stellar-sdk/pull/TODO))).
+* The `Server.getEvents()` method will correctly parse the event value as an `xdr.ScVal` rather than an `xdr.DiagnosticEvent` ([#TODO](https://github.com/stellar/js-stellar-sdk/pull/TODO)).
+
+
 ## v1.0.0-beta.4
 
 ### Fixed